+++ Changes.md
@@ -1,4 +1,9 @@
 
+## 1.0.4 - 2017-03-23
+
+- for outbound, find domain at hmail.todo.domain then hmail.domain.
+- noop: use es6 arrow functions
+
 ## 1.0.3 - 2017-03-09
 
 - add `enabled=false` flag for each limit type, defaults to off, matching the docs.